add filter functionality and only render products that are available
