+++ client/components/allProducts.js
@@ -1,9 +1,8 @@
 import React, {Component} from 'react';
-import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom';
 import {connect} from 'react-redux'
-import {fetchProducts} from '../store'
-import {Categories} from './'
+import {fetchProducts, fetchCategories} from '../store'
+import { withRouter } from 'react-router'
 
 
 /**
@@ -11,17 +10,52 @@ import {Categories} from './'
  */
 class AllProducts extends Component {
 
+  constructor (props) {
+    super(props);
+    this.state = {
+      //if the url has a category then set it to that category. If it does not then initialize state
+      currentCategory: props.match.params.category || ''
+    }
+    this.changeCategory = this.changeCategory.bind(this)
+  }
+
   componentDidMount () {
     this.props.loadProducts()
+    this.props.loadCategories()
+  }
+
+  changeCategory (categoryName) {
+    this.setState({currentCategory: categoryName.toLowerCase()})
+  }
+
+  componentWillReceiveProps (nextProps) {
+    //for rendering all products when the 'products' navlink is clicked when currentCategory is a category
+    const nextCategory = nextProps.match.params.category;
+    if (!nextCategory) {
+      this.changeCategory('');
+    }
   }
 
   render () {
+    //filter products if there is a current category
+    const products = (this.state.currentCategory ? this.props.allProducts.filter(product => {
+      return product.categories[0].name.toLowerCase() === this.state.currentCategory
+    }) : this.props.allProducts)
 
-    const products = this.props.allProducts;
+    const categories = this.props.categories;
+
 
     return (
       <div>
-        <Categories />
+      <div className="categories">
+      {
+        categories.map(category => {
+          return (
+            <Link to={`/products/category/${category.name.toLowerCase()}`} onClick={() => {this.changeCategory(category.name)}} className="category" key={category.id}>{category.name}</Link>
+          )
+        })
+      }
+    </div>
         <div className="product-group">
           {
             products.map(product => {
@@ -46,7 +80,8 @@ class AllProducts extends Component {
  */
 const mapState = (state) => {
   return {
-    allProducts: state.allProducts
+    allProducts: state.allProducts,
+    categories: state.categories
   }
 }
 
@@ -55,15 +90,13 @@ const mapDispatch = function (dispatch) {
     loadProducts: function () {
       const productsThunk = fetchProducts();
       dispatch(productsThunk);
+    },
+    loadCategories: function () {
+      const categoriesThunk = fetchCategories();
+      dispatch(categoriesThunk);
     }
   };
 }
 
-export default connect(mapState, mapDispatch)(AllProducts)
+export default connect(mapState, mapDispatch)(withRouter(AllProducts))
 
-// /**
-//  * PROP TYPES
-//  */
-// AllProducts.propTypes = {
-//   products: PropTypes.array
-// }
