+++ client/components/index.js
@@ -9,5 +9,5 @@ export {default as AllProducts} from './allProducts'
 export {Login, Signup} from './auth-form'
 export {default as Cart} from './cart'
 export {default as SingleProduct} from './SingleProduct'
-export {default as Categories} from './categories'
+
