+++ server/api/products.js
@@ -1,10 +1,10 @@
 const router = require('express').Router()
-const {Product} = require('../db/models')
+const {Product, Category} = require('../db/models')
 module.exports = router
 
 // returns all products
 router.get('/', (req, res, next) => {
-  Product.findAll()
+  Product.findAll({include: [{model: Category}]})
     .then(products => res.json(products))
     .catch(next)
 })
